move tests to fix linting issue
