+++ datahub-web-react/src/app/ingestV2/source/builder/RecipeForm/__tests__/common.test.ts
@@ -1,8 +1,20 @@
-import { setFieldValueOnRecipe, setListValuesOnRecipe } from '@app/ingestV2/source/builder/RecipeForm/common';
+import { describe, expect, it } from 'vitest';
+
+import {
+    getColumnProfilingCheckboxValue,
+    profilingEnabledFieldPath,
+    profilingTableLevelOnlyFieldPath,
+    removeStaleMetadataEnabledFieldPath,
+    setFieldValueOnRecipe,
+    setListValuesOnRecipe,
+    setRemoveStaleMetadataOnRecipe,
+    statefulIngestionEnabledFieldPath,
+    updateProfilingFields,
+} from '@app/ingestV2/source/builder/RecipeForm/common';
 
 describe('setFieldValueOnRecipe', () => {
     const accountIdFieldPath = 'source.config.account_id';
-    const profilingEnabledFieldPath = 'source.config.profiling.enabled';
+    const testProfilingEnabledFieldPath = 'source.config.profiling.enabled';
     const dottedFieldPath = ['source', 'config', 'profiling', 'enabled.test'];
 
     it('should set the field value on a recipe object when it was not defined', () => {
@@ -37,25 +49,25 @@ describe('setFieldValueOnRecipe', () => {
 
     it('should set the field value on a recipe object when it was not defined and has a parent', () => {
         const recipe = { source: { config: {} } };
-        const updatedRecipe = setFieldValueOnRecipe(recipe, true, profilingEnabledFieldPath);
+        const updatedRecipe = setFieldValueOnRecipe(recipe, true, testProfilingEnabledFieldPath);
         expect(updatedRecipe).toMatchObject({ source: { config: { profiling: { enabled: true } } } });
     });
 
     it('should update the field value on a recipe object when it was defined and has a parent', () => {
         const recipe = { source: { config: { profiling: { enabled: true } } } };
-        const updatedRecipe = setFieldValueOnRecipe(recipe, false, profilingEnabledFieldPath);
+        const updatedRecipe = setFieldValueOnRecipe(recipe, false, testProfilingEnabledFieldPath);
         expect(updatedRecipe).toMatchObject({ source: { config: { profiling: { enabled: false } } } });
     });
 
     it('should update the field value with a parent on a recipe without changing any other fields', () => {
         const recipe = { source: { config: { test: 'test' } } };
-        const updatedRecipe = setFieldValueOnRecipe(recipe, false, profilingEnabledFieldPath);
+        const updatedRecipe = setFieldValueOnRecipe(recipe, false, testProfilingEnabledFieldPath);
         expect(updatedRecipe).toMatchObject({ source: { config: { test: 'test', profiling: { enabled: false } } } });
     });
 
     it('should clear the field and its parent when passing in null and field is only child of parent', () => {
         const recipe = { source: { config: { test: 'test', profiling: { enabled: true } } } };
-        const updatedRecipe = setFieldValueOnRecipe(recipe, null, profilingEnabledFieldPath);
+        const updatedRecipe = setFieldValueOnRecipe(recipe, null, testProfilingEnabledFieldPath);
         expect(updatedRecipe).toMatchObject({ source: { config: { test: 'test' } } });
     });
 
@@ -139,5 +151,114 @@ describe('setListValuesOnRecipe', () => {
         const recipe = { source: { config: { table_pattern: { allow: ['*test_pattern'], deny: ['test_deny'] } } } };
         const updatedRecipe = setListValuesOnRecipe(recipe, [], tableAllowFieldPath);
         expect(updatedRecipe).toMatchObject({ source: { config: { table_pattern: { deny: ['test_deny'] } } } });
+    });
+});
+
+describe('updateProfilingFields', () => {
+    it('should enable table profiling only when table profiling is enabled and column profiling is disabled', () => {
+        const recipe = { source: { config: {} } };
+        const result = updateProfilingFields(recipe, true, false);
+
+        expect(result.source.config.profiling.enabled).toBe(true);
+        expect(result.source.config.profiling.profile_table_level_only).toBe(true);
+    });
+
+    it('should enable both table and column profiling when both are enabled', () => {
+        const recipe = { source: { config: {} } };
+        const result = updateProfilingFields(recipe, true, true);
+
+        expect(result.source.config.profiling.enabled).toBe(true);
+        expect(result.source.config.profiling.profile_table_level_only).toBe(false);
+    });
+
+    it('should disable profiling when table profiling is disabled', () => {
+        const recipe = { source: { config: {} } };
+        const result = updateProfilingFields(recipe, false, false);
+
+        expect(result.source.config.profiling.enabled).toBe(false);
+        expect(result.source.config.profiling.profile_table_level_only).toBeUndefined();
+    });
+
+    it('should disable profiling when table profiling is disabled regardless of column profiling', () => {
+        const recipe = { source: { config: {} } };
+        const result = updateProfilingFields(recipe, false, true);
+
+        expect(result.source.config.profiling.enabled).toBe(false);
+        expect(result.source.config.profiling.profile_table_level_only).toBeUndefined();
+    });
+});
+
+describe('getColumnProfilingCheckboxValue', () => {
+    it('should return true when profile_table_level_only is false', () => {
+        const recipe = {
+            source: {
+                config: {
+                    profiling: {
+                        profile_table_level_only: false,
+                    },
+                },
+            },
+        };
+        const result = getColumnProfilingCheckboxValue(recipe);
+
+        expect(result).toBe(true);
+    });
+
+    it('should return false when profile_table_level_only is true', () => {
+        const recipe = {
+            source: {
+                config: {
+                    profiling: {
+                        profile_table_level_only: true,
+                    },
+                },
+            },
+        };
+        const result = getColumnProfilingCheckboxValue(recipe);
+
+        expect(result).toBe(false);
+    });
+
+    it('should return false when profile_table_level_only is undefined', () => {
+        const recipe = { source: { config: {} } };
+        const result = getColumnProfilingCheckboxValue(recipe);
+
+        expect(result).toBe(false);
+    });
+});
+
+describe('setRemoveStaleMetadataOnRecipe', () => {
+    it('should enable both stateful ingestion and remove_stale_metadata when value is true', () => {
+        const recipe = { source: { config: {} } };
+        const result = setRemoveStaleMetadataOnRecipe(recipe, true);
+
+        expect(result.source.config.stateful_ingestion.enabled).toBe(true);
+        expect(result.source.config.stateful_ingestion.remove_stale_metadata).toBe(true);
+    });
+
+    it('should disable stateful ingestion and omit remove_stale_metadata when value is false', () => {
+        const recipe = {
+            source: {
+                config: {
+                    stateful_ingestion: {
+                        enabled: true,
+                        remove_stale_metadata: true,
+                    },
+                },
+            },
+        };
+        const result = setRemoveStaleMetadataOnRecipe(recipe, false);
+
+        expect(result.source.config.stateful_ingestion.enabled).toBe(false);
+        expect(result.source.config.stateful_ingestion.remove_stale_metadata).toBeUndefined();
+    });
+});
+
+describe('Field path constants', () => {
+    it('should have correct field paths', () => {
+        expect(profilingEnabledFieldPath).toBe('source.config.profiling.enabled');
+        expect(profilingTableLevelOnlyFieldPath).toBe('source.config.profiling.profile_table_level_only');
+        expect(statefulIngestionEnabledFieldPath).toBe('source.config.stateful_ingestion.enabled');
+        expect(removeStaleMetadataEnabledFieldPath).toBe('source.config.stateful_ingestion.remove_stale_metadata');
     });
 });
